+++ config/config.php
@@ -1,0 +1,12 @@
+<?php
+
+/***
+ *  Add Guzzle options (authentication, redirects, base urls, etc) here.
+ *  By default SSL verification is turned off to maintain backwards compatibility,
+ *  but you probably want to turn this on by removing the corresponding line.
+ */
+return [
+    'default-config' => [
+        'verify' => false,
+    ]
+];