+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `webhook` will be documented in this file
 
+## 1.0.1 - 2019-07-10
+
+- Adds compatibility up to Laravel 5.8
+- Allow for multiple URLs to be hit per notification by returning an array from `routeNotificationForWebhook` 
+- Adds configuration options for the internal Guzzle client
+
 ## 1.0.0 - 201X-XX-XX
 
 - initial release